add exception for all tap

+++ src/main/java/t3h/hostelmanagementsystem/entity/Category.java
@@ -29,7 +29,7 @@ public class Category {
     private String description;
 
     @Column(name = "status")
-    private Byte status = 1; // Default value: 1 (active)
+    private Integer status = 1; // Default value: 1 (active)
 
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
