+++ src/main/java/t3h/hostelmanagementsystem/entity/Hostel.java
@@ -45,7 +45,7 @@ public class Hostel {
     private List<Room> rooms;
 
     @Column(name = "status")
-    private Byte status = 1;
+    private Integer status = 1;
 
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
