+++ src/main/java/t3h/hostelmanagementsystem/entity/Room.java
@@ -49,7 +49,7 @@ public class Room {
     private String description;
 
     @Column(name = "status")
-    private Byte status = 1;
+    private Integer status = 1;
 
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt = LocalDateTime.now();
